Add files via upload
This is a haunted house prop controller. It plays two wave files, ambient and scare when an input is triggered. It can sequence the 8 outputs for a nearly unlimed period of time. The hard limit is the amount of ram. The sequence steps at a rate of 20 per second. The music is generated by the processor. Only outside requirement is the sound adapter https://www.pjrc.com/store/teensy3_audio.html